// src/app/components/merchandising/order-management/model/process-order-management.ts
export class ProcessOrderManagement {
  orderId: number;
  buyerId: number;
  seasonId: number;
  yearId: number;
  brandId: number;
  styleId: number;
  buyDate: string;
  orderNo: string;
  splitOrderNo: string;
  styleNo: string;
  landingPortName: string;
  destinationName: string;
  sizeValue: string;
  poDeliveryStatus: string;
  processOrderManagements: [];
}